move scripts that touch github api to it's designated folder in networks directory
